+++ R/list_datasets.R
@@ -13,11 +13,8 @@ list_datasets <- function() {
 
   data_sets <- tibble::tibble("name" = unlist(data_sets))
 
+
   return(data_sets)}
-
-
-
-