+++ NEWS.md
@@ -1,3 +1,5 @@
+# phsopendata (development version)
+
 # phsopendata 0.1.0 (2021-07-22)
 
 - Initial package release.